requirements.txt
# Core dependencies for LLM evaluation system
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0

# LLM providers
openai>=1.0.0
google-generativeai>=0.3.0
groq>=0.4.0
huggingface-hub>=0.16.0
transformers>=4.30.0
torch>=2.0.0

# RAG and vector search
faiss-cpu>=1.7.4
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Data processing and analysis
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
structlog>=23.0.0

# Security and validation
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Additional utilities
requests>=2.31.0
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0 